将垃圾数量选择 用户积分 从int类型转化为double类型

+++ src/main/java/com/work/recycle/entity/BaseOrder.java
@@ -16,7 +16,7 @@ public class BaseOrder {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-    private Integer score = 0;
+    private double score = 0;
 
     private Boolean receiveStatus = false;
